. proxy.sip, server.sip, genpy.cpp: added SetTimeoutSec for asynchronous request handler to support the timeout attribute in the method declarations. . actsvr.py, kasvr.py: fixed a bug that compared the status code of int type with the converted return value of uint32 type.

+++ python/tests/actcancel/actcsvr.py
@@ -47,7 +47,7 @@ class CActcServer:
         return listResp
 
     def LongWaitCleanup( self, ret, timerObj ) :
-        if np.uint32( ret ) == ErrorCode.ERROR_USER_CANCEL :
+        if ret == ErrorCode.ERROR_USER_CANCEL :
             self.DisableTimer( timerObj )
             print( "remove the timer",
                 "at user's request" );
